+++ tests/testthat/test-get_code_scale.R
@@ -2,7 +2,7 @@ context("test-get_code_scale")
 
 test_that("get_code_scale works", {
   
-  pal_fill_discrete <- get_code_scale(fill = "color", filltype = "categorical", params = list(palette = "Dark2"))
+  pal_fill_discrete <- get_code_scale(fill = "color", filltype = "discrete", params = list(palette = "Dark2"))
   expect_is(pal_fill_discrete, "character")
   expect_true(grepl("scale_fill_brewer", pal_fill_discrete))
   
@@ -10,10 +10,10 @@ test_that("get_code_scale works", {
   expect_is(pal_fill_continuous, "character")
   expect_true(grepl("scale_fill_distiller", pal_fill_continuous))
   
-  pal_fill_ggplot2 <- get_code_scale(fill = "color", filltype = "categorical", params = list(palette = "ggplot2"))
+  pal_fill_ggplot2 <- get_code_scale(fill = "color", filltype = "discrete", params = list(palette = "ggplot2"))
   expect_null(pal_fill_ggplot2)
   
-  pal_color_discrete <- get_code_scale(color = "color", colortype = "categorical", params = list(palette = "Dark2"))
+  pal_color_discrete <- get_code_scale(color = "color", colortype = "discrete", params = list(palette = "Dark2"))
   expect_is(pal_color_discrete, "character")
   expect_true(grepl("scale_color_brewer", pal_color_discrete))
   
@@ -21,7 +21,7 @@ test_that("get_code_scale works", {
   expect_is(pal_color_continuous, "character")
   expect_true(grepl("scale_color_distiller", pal_color_continuous))
   
-  pal_color_ggplot2 <- get_code_scale(color = "color", colortype = "categorical", params = list(palette = "ggplot2"))
+  pal_color_ggplot2 <- get_code_scale(color = "color", colortype = "discrete", params = list(palette = "ggplot2"))
   expect_null(pal_color_ggplot2)
 
 })