+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Module to coerce a variable to a different type.
 * Scroll when data have a lot of variables.
-* Prevent filter for categorical variables with over 100 unique elements.
+* Prevent filter for discrete variables with over 100 unique elements.
 
 
 # esquisse 0.1.0